# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TechPulse AI API",
    "version": "1.7.0",
    "description": "Canonical HTTP contract for the Vietnamese-first TechPulse AI MVP. Browser API access is same-origin only; external content is untrusted data; no response contains source full text or media binary."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin API. Production sends no CORS allow-origin/credentials headers; any future exception requires an exact reviewed scheme+host allowlist."
    }
  ],
...
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "registerUser",
        "x-persistence": "mongo",
        "summary": "Register a user account",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/BrowserOriginHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account and session created",
            "headers": {
              "Set-Cookie": {
                "$ref": "#/components/headers/SetSessionCookie"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-techpulse_session",
        "description": "Opaque server-side session token. The cookie is host-only and always emitted with Secure, HttpOnly, Path=/ and SameSite=Lax; Domain is forbidden. Every cookie-authenticated mutation also passes the exact same-origin browser Origin gate and CSRF validation."
      },
      "cronBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Vercel CRON_SECRET; never accepted from a browser session."
      }
    },
    "headers": {
      "SetSessionCookie": {
        "description": "Exactly __Host-techpulse_session=<opaque>; Path=/; Secure; HttpOnly; SameSite=Lax. Domain is absent and Max-Age never exceeds the session absolute expiry.",
        "schema": {
          "type": "string"
        }
      },
...
      "PrivateNoStore": {
        "description": "Authenticated/session-bearing responses must not be stored by browser or shared caches.",
        "schema": {
          "type": "string",
          "const": "no-store, private"
        }
      },
...
    },
    "parameters": {
      "BrowserOriginHeader": {
        "name": "Origin",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uri",
          "maxLength": 2048
        },
        "description": "Browser origin. Production requires an exact normalized scheme+host+effective-port match with the configured public application origin; suffix, wildcard and reflected matching are forbidden."
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "bad_request",
                "message": "Malformed request",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
...
      "Conflict": {
        "description": "State, duplicate, lease, or idempotency conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "conflict",
                "message": "State, duplicate, lease, or idempotency conflict",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Semantic validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "validation_error",
                "message": "Semantic validation failed",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "headers": {
          "Retry-After": {
            "description": "Seconds until retry is permitted",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "rate_limit_exceeded",
                "message": "Rate limit exceeded",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "Request target or JSON payload exceeds the global ingress limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "payload_too_large",
                "message": "Request exceeds the allowed size",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "JSON route requires application/json with identity content encoding",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "unsupported_media_type",
                "message": "Unsupported request content type or encoding",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected internal failure",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "internal_error",
                "message": "Unexpected internal failure",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Required dependency is temporarily unavailable",
        "headers": {
          "Retry-After": {
            "description": "Seconds until a retry may succeed",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "service_unavailable",
                "message": "Required dependency is temporarily unavailable",
                "requestId": "req_01HXYZ"
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "UserStatus": {
        "type": "string",
        "enum": [
          "active",
          "suspended",
          "deletion-pending",
          "deleted"
        ]
      },
...
      "User": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "email",
          "role",
          "status",
          "topicPreferences",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "topicPreferences": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true,
            "maxItems": 20
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 10,
            "maxLength": 128
          }
        }
      },
...
      "AuthData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "user",
          "csrfToken"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "csrfToken": {
            "type": "string",
            "minLength": 32,
            "description": "Send in X-CSRF-Token for state-changing session-authenticated requests."
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuthData"
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message",
          "code"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "bad_request",
          "validation_error",
          "unauthorized",
          "csrf_invalid",
          "forbidden",
          "not_found",
          "conflict",
          "idempotency_mismatch",
          "invalid_state_transition",
          "source_policy_blocked",
          "rate_limit_exceeded",
          "payload_too_large",
          "unsupported_media_type",
          "provider_unavailable",
          "database_unavailable",
          "service_unavailable",
          "internal_error"
        ],
        "description": "Stable machine-readable taxonomy. Operations may document a narrower subset; clients must not branch on message text."
      },
      "ErrorObject": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorObject"
          }
        }
      }
    }
  }
}